multer

+++ NodeJSProj/controllers/admin.js
@@ -1,11 +1,11 @@
 const Product = require("../models/product");
-const { validationResult } = require('express-validator')
-const mongoose = require('mongoose')
+const { validationResult } = require("express-validator");
+const mongoose = require("mongoose");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
-
-  if(!req.session.isLoggedIn) {
-    return res.redirect('/login')
+  if (!req.session.isLoggedIn) {
+    return res.redirect("/login");
   }
 
   res.render("admin/edit-product", {
@@ -14,34 +14,51 @@ exports.getAddProduct = (req, res, next) => {
     editing: false,
     hasError: false,
     errorMessage: null,
-    validationErrors: []
+    validationErrors: [],
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
-  const errors = validationResult(req)
-
-  if(!errors.isEmpty()) {
-   return res.status(422).render("admin/edit-product", {
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/add-product",
       editing: false,
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
-        description: description
+        description: description,
       },
-      errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      errorMessage: "Attached file is not an image",
+      validationErrors: [],
     });
   }
+
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+    });
+  }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     // _id: new mongoose.Types.ObjectId('619d0ec275d30a9130ac2a32'),
@@ -58,7 +75,7 @@ exports.postAddProduct = (req, res, next) => {
       console.log("Created product!");
       res.redirect("/admin/products");
     })
-    .catch(err => {
+    .catch((err) => {
       // return res.status(500).render("admin/edit-product", {
       //   pageTitle: "Add Product",
       //   path: "/admin/add-product",
@@ -73,10 +90,10 @@ exports.postAddProduct = (req, res, next) => {
       //   errorMessage: 'Database operation failed, please try again.',
       //   validationErrors: []
       // });
-      const error = new Error(err)
+      const error = new Error(err);
       error.httpStatusCode = 500;
 
-      return next(error)
+      return next(error);
     });
 };
 
@@ -99,14 +116,14 @@ exports.getEditProduct = (req, res, next) => {
         product: product,
         hasError: false,
         errorMessage: null,
-        validationErrors: []
+        validationErrors: [],
       });
     })
-    .catch(err => {
-      const error = new Error(err)
+    .catch((err) => {
+      const error = new Error(err);
       error.httpStatusCode = 500;
-      
-      return next(error)
+
+      return next(error);
     });
 };
 
@@ -114,79 +131,88 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
-  const errors = validationResult(req)
+  const errors = validationResult(req);
 
-  if(!errors.isEmpty()) {
-   return res.status(422).render("admin/edit-product", {
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
       editing: true,
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
-        _id: prodId
+        _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
   Product.findById(prodId)
     .then((product) => {
-      if(product.userId.toString() !== req.user._id.toString()) {
-        return res.redirect('/')
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
-      return product.save()
-      .then(() => {
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl); // fire and forget
+        product.imageUrl = image.path;
+      }
+      return product.save().then(() => {
         console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
       });
     })
-    .catch(err => {
-      const error = new Error(err)
+    .catch((err) => {
+      const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error)
+      return next(error);
     });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId: req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then((products) => {
       res.render("admin/products", {
         prods: products,
         pageTitle: "Admin Products",
-        path: "/admin/products"
+        path: "/admin/products",
       });
     })
-    .catch(err => {
-      const error = new Error(err)
+    .catch((err) => {
+      const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error)
+      return next(error);
     });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      
+      fileHelper.deleteFile(product.imageUrl); // fire and forget
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT!");
       res.redirect("/admin/products");
     })
-    .catch(err => {
-      const error = new Error(err)
+    .catch((err) => {
+      const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error)
+      return next(error);
     });
 };
